requirements.txt
requests>=2.31.0
tenacity>=8.2.3
pandas>=2.0.3
sqlalchemy>=2.0.19
psycopg2-binary>=2.9.7
python-dotenv>=1.0.0
dbt-postgres>=1.5.0
great-expectations>=0.17.11
pytest>=7.4.0
pytest-cov>=4.1.0
